ルーレットの当たったマスの計算方法を修正

+++ simpleRulette/View/ViewController/RuletteViewController.swift
@@ -109,13 +109,13 @@ class RuletteViewController: UIViewController, UIGestureRecognizerDelegate {
                 }
                 
                 // 360° = 100% とした時の割合
-                let per = Int((testAngle) / 3.60)
+                let per = (testAngle / 3.60)
                 
                 var itemValue = 0
                 var count = 0
                 for item in items {
                     itemValue += 100 / items.count
-                    if per < itemValue || count == items.count {
+                    if per < Double(itemValue) || count == items.count {
                         resultLabel.text = "結果は\(item)です"
                         return
                     }
